+++ OPHD/UI/StructureListBox.cpp
@@ -128,7 +128,7 @@ void StructureListBox::currentSelection(Structure* structure)
 		StructureListBoxItem* item = static_cast<StructureListBoxItem*>(mItems[i]);
 		if (item->structure == structure)
 		{
-			setSelection(static_cast<int>(i));
+			setSelection(i);
 			return;
 		}
 	}
@@ -168,8 +168,8 @@ void StructureListBox::update()
 		drawItem(renderer, *static_cast<StructureListBoxItem*>(mItems[i]),
 			positionX(),
 			positionY() + (static_cast<int>(i) * LIST_ITEM_HEIGHT),
-			item_width(),
-			draw_offset(),
+			static_cast<int>(item_width()),
+			static_cast<int>(draw_offset()),
 			i == currentSelection());
 	}
